notification html and moving code
