+++ CRUD/admin.py
@@ -2,5 +2,8 @@ from django.contrib import admin
 from .models import AdjunctFacultyMember
 
 
+
+
 # Register your models here.
 admin.site.register(AdjunctFacultyMember)
+
