+++ CRUD/forms.py
@@ -2,7 +2,7 @@ from django import forms
 from django.forms.models import ModelForm
 from .models import AdjunctFacultyMember
 from .models import Classes
-from .models import Notification
+# from .models import Notification
 
 
 class DateInput(forms.DateInput):
@@ -31,6 +31,7 @@ class ClassForm(forms.ModelForm):
     fields = 'class_name'
 
 
-class NotificationForm(forms.ModelForm):
-    model = Notification
-    fields = 'message', 'date'
+# class NotificationForm(forms.ModelForm):
+#     class Meta:
+#         model = Notification
+#         fields = 'message', 'date'
