+++ Notifications/forms.py
@@ -1,0 +1,8 @@
+from django import forms
+from django.forms.models import ModelForm
+from .models import Notification
+
+class NotificationForm(forms.ModelForm):
+    class Meta:
+        model = Notification
+        fields = 'message', 'date'
